feat: converted contrast-colors to es6 modules
