+++ packages/contrast-colors/.eslintrc.json
@@ -1,7 +1,7 @@
 {
     "env": {
         "browser": true,
-        "commonjs": true
+        "es6": true
     },
     "extends": [
         "airbnb-base"
