+++ packages/contrast-colors/test/ratioName.test.mjs
@@ -10,16 +10,16 @@ governing permissions and limitations under the License.
 */
 /* global test, expect */
 
-const { minPositive } = require('../index');
+import { ratioName } from "../index";
 
-test('should return 1', () => {
-  const result = minPositive([1, 1.2, 1.4, 2, 3, 4.5, 6, 8, 12, 21]);
+test('should output 10 numbers incremented by 100', () => {
+  const theme = ratioName([1, 1.2, 1.4, 2, 3, 4.5, 6, 8, 12, 21]);
 
-  expect(result).toEqual(1);
+  expect(theme).toEqual([100, 200, 300, 400, 500, 600, 700, 800, 900, 1000]);
 });
 
-test('should return 2', () => {
-  const result = minPositive([-3, -2, -1.2, 2, 3, 4.5, 6, 8, 12, 21]);
+test('should output 10 numbers with first at 50', () => {
+  const theme = ratioName([-1.5, 1.2, 1.4, 2, 3, 4.5, 6, 8, 12, 21]);
 
-  expect(result).toEqual(2);
+  expect(theme).toEqual([50, 100, 200, 300, 400, 500, 600, 700, 800, 900]);
 });
