+++ packages/contrast-colors/test/theme.test.mjs
@@ -9,7 +9,7 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 /* global test, expect */
-const { Theme, Color, BackgroundColor } = require('../index');
+import { Theme, Color, BackgroundColor } from "../index";
 
 test('should generate theme for three colors', () => {
   const gray = new BackgroundColor({ name: 'gray', colorKeys: ['#cacaca', '#323232'], colorspace: 'HSL', ratios: [1, 1.2, 1.4, 2, 3, 4.5, 6, 8, 12, 21] });
