--- packages/contrast-colors/wrapper.mjs
@@ -1,13 +1,0 @@
-import contrastColors from './index.js';
-
-export const createScale = contrastColors.createScale;
-export const luminance = contrastColors.luminance;
-export const contrast = contrastColors.contrast;
-export const binarySearch = contrastColors.binarySearch;
-export const minPositive = contrastColors.minPositive;
-export const ratioName = contrastColors.ratioName;
-export const BackgroundColor = contrastColors.BackgroundColor;
-export const Color = contrastColors.Color;
-export const Theme = contrastColors.Theme;
-
-export default contrastColors;
